move os.start to runtime.start
